Truncated decimal points on loading screens.

+++ core/src/com/majalis/screens/LoadScreen.java
@@ -34,7 +34,7 @@ public class LoadScreen extends AbstractScreen {
 		batch.begin();
 		if (!assetManager.update() || clocktick < 25){
 			font.draw(batch, String.valueOf(clocktick++), 1675, 500);
-			font.draw(batch, "Loading: " + (assetManager.getProgress() * 100) + "%", 1125, 750);
+			font.draw(batch, "Loading: " + (int)(assetManager.getProgress() * 100) + "%", 1125, 750);
 		}	
 		else {
 			showScreen(screenRequest);
